+++ TinkoffASDKCore/TinkoffASDKCoreTests/IntegrationTests.swift
@@ -27,7 +27,6 @@ class IntegrationTests: XCTestCase {
 
     override func setUp() {
         let credential = AcquiringSdkCredential(terminalKey: StageTestData.terminalKey,
-                                                password: StageTestData.terminalPassword,
                                                 publicKey: StageTestData.testPublicKey)
 
         let acquiringSDKConfiguration = AcquiringSdkConfiguration(credential: credential)