+++ TinkoffASDKUI/TinkoffASDKUI/PullableContainer/Views/PullableContainerHeader.swift
@@ -17,6 +17,7 @@
 //  limitations under the License.
 //
 
+import UIKit
 
 final class PullableContainerHeader: UIView {