+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * [MIC-6194] SPB Feature is False by default in AcquiringViewConfiguration.FeaturesOptions
 * [MIC-6292] Swift Gen integration for TinkoffASDKUI module
+* [MIC-6292] Swift Package Support & buid Package.swift step for CI/CD
 
 ### Added